Use Loofah to scrub HTML attachments
We can't use `#sanitize` here because it operates on a Loofah fragment
instead of a loofah document [1]. This results in the `<head>` and
`<body>` tags getting stripped and returning an invalid HTML page.

With Loofah's built in `:prune` scrubber we retain the old behaviour of
stripping out script tags.

`:prune` removes unknown/unsafe tags and their contents (including their
subtrees):

    unsafe_html = "ohai! <div>div is safe</div> <foo>but foo is
    <b>not</b></foo>"

    Loofah.fragment(unsafe_html).scrub!(:prune)
    # => "ohai! <div>div is safe</div> "

* Adds a `DOCTYPE` to the fixture file so that Nokogiri doesn't insert a
  HTML 4 `DOCTYPE` automatically, making comparison in the spec uglier
* Nokogiri also adds a `meta` tag to the output. Not much we can do
  about this: https://github.com/sparklemotion/nokogiri/issues/1008

[1] https://github.com/flavorjones/loofah#side-note-fragments-vs-documents