Try a new strategy for MailServerLog.set_delivery_status
For some reason (possibly caused by the serializer updating
delivery_status?) `set_delivery_status` is being treated as a dirty
write which means that Rails tries to access the old (invalid) attribute
value on calling write_attribute to update the model's `changes` hash,
which unintentionally re-raises the rescued error:

https://apidock.com/rails/v4.2.7/ActiveRecord/AttributeMethods/Dirty/write_attribute

To work around this, this commit introduces an optional `force` param
for `set_delivery_status`. If set, it will use `raw_write_attribute` to
blindly set the new value then calls `save_changed_attribute` to add an
entry to the `changes` hash so that `save` stores the value otherwise
`save` is a noop and the database will still have the old value, giving
the appearance of a silent failure condition.

(If/when we remove this, we may have to copy this approach - unless by
then we've moved onto a new Rails which does things differently - to the
`temp:cache_delivery_status` task as it hits the same error when calling
`update_attributes!` if the original value is invalid.)