Override the user name param with the attribute value before update
Fixes broken spec:

  ./spec/integration/admin_user_controller_edit_spec.rb:19

Since upgrading to Rails 4.2, the value for User#name for a banned
user is always supplied with the "(Account suspended)" suffix so
when updating user account details via a form, the supplied params
include "Name (Account suspended)" which causes the name and url_name
to be updated in the database.

The change that caused this is
https://github.com/rails/rails/commit/3e74db6361cf61434fa47f44fcbccf50ce02bad8.

At the point of rendering the form `value_came_from_user?(object)`
returns `false` so we end up with the suffix applied.