Add Pro User Signup action stub
Handle creation of a Pro user and rendering the “confirm your account”
page for the pro context.

I haven’t created a form here as we’ll be rendering it in the yet-to-be-
created SessionsController#new action.

We’ll add it in when we have a sessions controller, or use the existing
sign-up form.