WIP Require id in CSV import files
See CSV in https://github.com/mysociety/alaveteli/issues/4078 – doesn't
work.

Adding an `id` field with blank values allows the CSV to be imported
correctly.

I _think_ the changes made here are fine, but the spec failures need to
be looked at in more detail. I think what's happening is that the output
is just showing the newly added `id` attribute with the persisted value.