Add stripe-ruby-mock
This mocks Stripe calls so that we don’t hit their real API in tests.

This is a bit of a risky dependency, as it doesn’t sound like the
author has the capacity to maintain the gem at the moment [1].

There’s thoughtbot’s fake_stripe [2] gem, but that doesn’t have helpers
to allow you to create your own mock plans etc.

There’s also vcr [3], but this would require us to configure stripe API
keys for the test env which is a bit of a hassle at the moment.

So all in all I figured stripe-ruby-mock is the best solution
_for now_, but will need re-assessing soon.

[1] https://github.com/rebelidealist/stripe-ruby-mock#request-looking-for-more-core-contributors
[2] https://github.com/thoughtbot/fake_stripe
[3] railscasts.com/episodes/291-testing-with-vcr