Add basic handling of duplicate plan subscription
If a `User` has a `ProAccount`, we can assume that they're already
signed up to the 'pro' plan, as its the only one we offer right now.

This is much easier than checking Stripe for plans, but we'll need to
do this when we offer more than the standard pro plan.