squash! Add basic handling of duplicate plan subscription
Actually have to look at Stripe subscriptions because of the way we'll
want to handle failed charges later.

In the case of a failed charge, we want to redirect back here with an
error message. By then the Stripe customer has already been created, so
to prevent a duplicate customer we'll need to retrieve them through
their Stripe customer ID. This requires having a `ProAccount` with a
persisted `customer_id`, so preventing access to this page by
`ProAccount` presence alone will not work.