Display applied discount coupons on account page
If a user has an applied discount coupon, it should show up on their
account page and the amount should be deducted from their next payment
information so that it doesn't appear we are going to charge them the
full amount.

The assumption being made here is that discount coupons will be manually
applied to subscriptions via the Stripe dashboard rather than redeemed
through the site. Also that any coupon that's applied to the account
and marked as valid is fine to use - observation suggest that one-off
coupons are automatically deleted from subscriptions by the Stripe
service so hopefully there are no other checks needed here.

NB - this just calculates a for-display value, there is no effect on the
amount charged by Stripe.