Run problematic query through #find_by_sql
This query has started to run indefinitely on WDTK. It locks a passenger
process, so eventually results in a queue full error.

The query itself runs fine (in an SQL console), but for whatever reason
running it through arel just hangs.

Running it through `#find_by_sql` seems to work okay in the console.