Add UI to cancel subscription
Cancells the current Stripe subscription at the period end. Note that
clicking this does **not** remove any roles assigned to the `User`, as
they still need them to make use of the features in the remaining
subscription period.

We'll need to handle Stripe webhooks [1] in order to actually remove
features.

[1] https://stripe.com/docs/subscriptions/webhooks