Use UserPhaseCount to count categorised requests in the ToDoList
Use the user's phase_count for the count value of each named phase that
is handled by the ToDoList rather than having a database query for each.

Assumes that there are up to date summaries available but this seems
fair as it appears on the same page as the _projects sidebar which
relies on the same data.