Make _request_listing_single same as _request_listing_via_event
Going to use this for showing private requests on the user profile
(private requests are not indexed, so will be pulled out differently),
but the display is slightly different to _request_listing_via_event.

`_request_listing_single` is only used in the request game, so should be
fine to make these the same.