Prevent ProAccount#active? throwing an error without stripe customer id
Possibly a bit of an edge case but one that people with existing test
data are likely to hit. (We are planning to create Stripe accounts for
existing accounts before we switch this feature on in production so
hopefully this just applies to test data.)

The signup process is halted at the Plans#show controller with an
"undefined method subscriptions for nil::NilClass" error (attempting to
call active? for current_user.pro_account) if the user has a pro_account
record but no stored Stripe customer id