Don't run service command as app user
This fails on Debian Stretch (and probably any other systemd-based OS)
when run as an app user as the service command just passes this to
systemctl which needs root privileges.

For now, just call the script directly as its the easiest thing to do to
fix the problem, and we do similar in the Capistrano deployment config.