Set `ip_spoofing_check` to false
An 'IP spoofing attack?!' error is triggered by a conflict between
the CLIENT_IP and X_FORWARDED_FOR headers. When using the example
nginx config, nginx should be setting X_FORWARDED_FOR which is used in
preference to CLIENT_IP when the spoofing check is disabled. So this
setting should just prevent false positive errors when requests have
a CLIENT_IP header set.