Use Xapian results to set page count
`info_requests.size` exposes a rough idea of the amount of private
requests a pro user has made.

Since private requests are not indexed by Xapian, the
`matches_estimated` will provide the correct pagination.

Also rewords heading to convey result count is approximated as per
https://github.com/mysociety/alaveteli/pull/4052.

This isn't the best of wording, but its clearer than trying to jam
"approximately" in to the rest of the text.

Fixes https://github.com/mysociety/alaveteli-professional/issues/396.