Better handle a logged in user logging in or signing up
We check authenticity tokens for logged in users as a CSRF
protection. We don't check or generate tokens for non
logged in users so that we can strip cookies and cache
pages in varnish and serve them up to any non logged in
user.

If a user who is logged in tries to log in again or sign up
with a form generated before they logged in, the absence of
authenticity token would cause an `InvalidAuthenticityToken`
error.

We recently switched to handling these by raising an
exception rather than nullifying their session, so we
have greater visibility of the fact that users who
are already logged in are logging in again or signing
up.

I don't think showing an error page (or indeed
nullifying their session) is the best thing to do
here, and I also think it's safe to remove the CSRF
protection on these actions as we're not authorising
any action based on the session contents, only on the
form submitted.

I think the least surprising thing for the user here
is to behave as if they're not already logged in
(as that's what they apparently believe), so for
logging in, to log them in if they submit the right
credentials, and log them out if they submit the wrong
ones. For signing up, we handle the signup as we
would if they weren't already logged in.