Pass-through signin form if already logged in
We shouldn't render the signin form if a user is already logged in. Its
confusing and results in weird behaviour.

For example, if a user is signed in, receives a "You have a new response
" email and follows the URL within the email, they shouldn't need to
sign in again before being redirected to their new response, as the
session already exists.

Fixes https://github.com/mysociety/alaveteli-professional/issues/288
(and also fixes this for non-pro users).

I've currently disabled failing specs for redirecting to external URLs.

The weirdness of our custom error handling solution means these return a
500 status rather than a 404. The 500 is caused by the exception
handling raising the `RouteNotFound`.

We use this behaviour in `do_post_redirect` [1] so I'm confident it
solves the problem, and can be verified in the browser.

I've expanded on what I think is happening in
https://github.com/mysociety/alaveteli/issues/4353.

I've also forced setting of `consider_all_requests_local` to prevent
spec failures on < Ruby 2.3. I'm not sure if this is a Ruby version
issue, or whether the pass on Ruby 2.3 was just a lucky run where there
was no interaction with other specs that caused a failure.