fixup! Move confirmation action to the confirmation controller
This was added in f4e985074ee7a6a4d651bc92dc9d30e0889b5c92 and
ea1eec779de8221fdf9ff7cac56bc4fd2b181e7b so that we could redirect to
the no cookies page if the post redirect wasn't in the session.

We're not doing that now, so removing.