Reduce usage of auto-login links from mailers
Auto-login URLs are a security problem, as users sometimes forward
emails or tweet the confirmation URLs, so others can then log in as the
user.

We opted to remove these for Pro users given the sensitivity of
requests, and given the confusion it creates for regular users (and
additional maintenance for developers), this removes them entirely.

In some cases, we don't automatically get a sign-in form if we try to
complete the action, so instead we pass through the sign-in form. If the
user is already signed in, they'll get redirected without noticing. If
they need to sign in, they'll get redirect after that.

Some actions still generate auto-login links:

* Confirming your account
* Changing your password
* Changing your email address
* Authority upload URLs

At this point in time, I'm not sure what to do with these, but there's
much less likelihood that users will forward these emails, so we can
deal with these later.

`NotificationMailer#daily_summary` makes use of the hacky
`batch_request_id_mappings` as we can't specify the IDs of the requests
belonging to a batch.

* Remove auto-login from RequestMailer#overdue_alert
* Remove auto-login from RequestMailer#very_overdue_alert
* Remove auto-login from RequestMailer#new_response_reminder_alert
* Remove auto-login from RequestMailer#not_clarified_alert
* Remove auto-login from InfoRequestBatchMailer#batch_sent
* Remove auto-login from NotificationMailer#overdue_notification
* Remove auto-login from NotificationMailer#very_overdue_notification
* Remove auto-login from TrackMailer#event_digest
* Remove auto-login from NotificationMailer#daily_summary

Fixes https://github.com/mysociety/alaveteli-professional/issues/325
Fixes https://github.com/mysociety/alaveteli/issues/1324