Manually send a GA virtual pageview when a new Pro user signs up
Rather than using a visitable URL which could record force positives for
Goal completions, this manually sends a virtual  pageview that matches
the Google Analytics Goal Destination (this will override the standard
GA pageview code in the default layout template).

This will allow us to count the pageview as normal without inflating the
reported number of pageviews and still use all of the reporting features
associated with Goal Steps and Funnel Visualisation reports.