Set IDs for factory data
To check that `NotificationMailer#daily_summary` is doing the right
thing, we assert that the mail body is the same as a fixture template.

The problem with this is that when we remove the `PostRedirect`-based
URLs, we're going to need the ID of the request to link to.

We need this to be static, otherwise we'll get failures after the
requests are created with dynamically allocated ID numbers.