Duplicate marketing feature list to /pro/pricing
The page just feels a bit bare without it. Its a nice reminder of the
features, and gives us chance to make the screenshots a bit bigger.

I haven't extracted these in to partials because of the screenshot issue
(we'd have to do more CSS) and its more of a temporary page for now
anyway.