Download RawEmail as .eml instead of .txt
I almost _always_ want to open the email in Apple Mail so that I can
extract attachments that we had trouble parsing.

To do that, I download the raw email, _rename the file extension_ and
then open in Apple Mail. This adds a really annoying step to the task at
hand.

This commit downloads the email with a `.eml` extension so that
double-clicking the file opens it in Apple Mail (or your chosen mail
client) automatically.

Its still plain text, so nothing needs to be done when operating on the
file with command line tools.

Added a custom mime type as per
http://guides.rubyonrails.org/v4.2.10/action_controller_overview.html#restful-downloads