Make Trusty work with Ruby 2.x
Trusty has a known bug which has made update-alternatives inompatible
with the available ruby2.0 package, preventing the version from being
changed cleanly, so we can't rely on using that to set ruby 2.
See: https://bugs.launchpad.net/ubuntu/+source/ruby2.0/+bug/1310292

This installs Ruby 2.1.x using the Brightbox packaged version.
(Currently this installs 2.1.9 which gets reported as 2.1.5 by
our install script)

The alternative approach of manually creating symlinks for ruby and gem
to point to the newer version - as described here
https://bugs.launchpad.net/ubuntu/+source/ruby2.0/+bug/1310292/comments/20
- seems to work when setting up a Vagrant box install from scratch but
is likely to have unintended consequences for anyone using an existing
box and/or a using a ruby environment manager so I am reluctant to
recommend it.