Minor change to user drop down positioning
To ensure the whole button behaves better when the navigation menu needs
to overflow. Previously, due to the way the arrow icon was positioned it
didn't follow the button when it overflowed to a new line.

Themes may need a tweak for this fix if they have overridden the _right_
value of _.js-loaded .account-link:after_