squash! Added Binstubs for supported vagrant OSs
Alternative approach for multi-OS Vagrant.

Instead of defining a new "name" setting, just use the existing "os"
key.

The problem with suggesting this initially was that it would force a
rebuild of the "default" VM. This defines the default VM in addition to
a multi-machine [1] VM for each supported OS.

The benefit of this approach is that each appears in `vagrant status`
output:

    $ vagrant status
    Current machine states:

    default                   running (virtualbox)
    trusty64                  not created (virtualbox)
    jessie64                  not created (virtualbox)
    stretch64                 not created (virtualbox)

    This environment represents multiple VMs. The VMs are all listed
    above with their current state. For more information about a specific
    VM, run `vagrant status NAME`.

The binstubs now run `vagrant CMD VM_NAME`:

    $ bin/vagrant-stretch status
    Current machine states:

    stretch64                 not created (virtualbox)

    The environment has not yet been created. Run `vagrant up` to
    create the environment. If a machine is not created, only the
    default provider will be shown. So if a provider is not listed,
    then the machine is not created for that environment.

Each VM (other than the primary [2]) is set to not start unless
explicitly requested [3].

[1] https://www.vagrantup.com/docs/multi-machine/
[2] https://www.vagrantup.com/docs/multi-machine/#specifying-a-primary-machine
[3] https://www.vagrantup.com/docs/multi-machine/#autostart-machines