Don't call out to Stripe::Customer.create with pro_pricing disabled
As there is no rescue for the begin block in
ProAccount#set_stripe_customer_id, an error will be raised if this
method is called when STRIPE_SECRET_KEY has not been set. As this should
raise an error if pro_pricing is enabled, adding a guard to trigger an
early return when pro_pricing is deliberately disabled seems like the
simplest option, avoiding an extra call out to the Stripe API if none is
needed.