Show public body notes in the batch builder
MVP of adding important notes to the batch builder. Here we just show
any notes if the authority has any notes. We want to be more selective
about which notes we show, but that will come in a future commit.

The notes are closed by default in the search results pane. When a user
adds them, the notes are automatically expanded in the selected
authorities pane.