Show public body notes in the batch builder
MVP of adding important notes to the batch builder. Here we just show
any notes if the authority has any notes. We want to be more selective
about which notes we show, but that will come in a future commit.

Originally I wanted to automatically open the `<details>` when added to
the selected pane, but this opens _all_ `<details>`, even if you've
closed previously added authorities.

This is the next best solution. When you add an authority to the
selection pane the details are closed. You can manually open the
details. If you then add another authority, any open details in the
selection pane get closed. This seems like acceptable behaviour, as
you'll have already read the notes if you're going back for more
authorities.