Use low BCrypt cost in specs
We don't need to do really expensive hashing while running specs, so
this should reduce the spec time a bit.

The controller specs seemed to go from ~17 minutes to ~12 minutes with
this change when I tried it, but I was doing other things at the time so
it wasn't a totally scientific test.