Log events in InfoRequest.stop_new_responses_on_old_requests
We log changes to `allow_new_responses_from` when edited through the
admin interface, but not when we run
`InfoRequest.stop_new_responses_on_old_requests` through our daily cron
job.

Its quite important to know when we change response behaviour, so
logging an event exposes this nicely in the admin interface of the
request page.

There are a few compromises in this change.

1. Performance is reduced slightly as we're using `find_in_batches` over
   a direct SQL `UPDATE`. This is necessary so that we can call
   `log_event` on each `InfoRequest` we update.
2. We log the event inline. Ideally we'd use an `after_update` callback,
   but there's no good way of _not_ touching `updated_at` when going
   through the ActiveRecord stack. Rails 5 introduces a `touch: false`
   option to `save` [1], so we can probably use that later on.
3. There's duplucation in the `find_in_batches` block. If we handled
   event logging in a callback, we could remove this duplication.

Fixes https://github.com/mysociety/alaveteli/issues/4531.

[1]
http://api.rubyonrails.org/classes/ActiveRecord/Persistence.html#method-i-save