Add notes about password length.
Forgot this was now a limitation when signing up with a test user.

Hopefully prevent a few people getting frustrated after entering a
password thinking its okay, and then getting told its not.

I was going to put this in a placeholder as I saw on
https://highrisehq.com, but the `form_item_note` is more consistent with
the other form notes.