Hack the authorities search to use the N weighting code
Crudely introduces a user_query param to `perform_search` so that we can
pass a pre-constructed Xapian::Query to Xapian::Search rather than a
string.

The downside of this approach - aside from adding a 7th param to the
perform_search method - is that it makes the search controller
responsible for formatting the search string and constructing the query.