Retain draft when switching batch mode
Quick fix for preventing the authority selection getting lost if a user
switches mode. Ideally they should be able to add authorities from a
category, then switch to search mode to refine the selection.

From @gbp in this approach:

> This looks like it would work, can't see why it would break
>
> On the code design; I think generally the JS for the batch builder has
been designed to be based around components (EG. body-list,
write-button, ...) for different part of the UI, which communicate
actions using events. If we want to stay true to this then create a new
component for the tabbed nav, listen to the same events and put this
code in there - although this might be overkill here especially as we
hopefully will be able to revisit and improve once we have the ability
to use a frontend framework post Rails 5.