Prioritise Xapian N terms for PublicBody
This boosts the WDF increment for terms with the "N" prefix – namely
`PublicBody#name`, prioritising direct matches when a user searches for
a public body by name.

Xapian gives the option of doing this either at index time or at search
time [1].

@lizconlan attempted the latter [2], but it resulted in having to jam
lots of extra logic in to our search methods. Given indexing WDTK's
public bodies only takes around 5 minutes (for ~23K records) it seems
more efficient to do this at index time. It also results in much less
code changed.

The actual weighting is quite strange. The example [1] seems to suggest
that you want a lower termpos and a higher WDF increment for a
prioritised term. I'm not entirely sure what the interplay between the
term position and WDF increment is. For future reference, I'm including
links to `add_posting` [3] and `add_term` [4].

Note that the specs assert on `PublicBody#name` instead of comparing
object equality because the rspec diff isn't very helpful now that name
isn't an attribute on `PublicBody` directly (its on
`PublicBody::Translation`).

Aside: Ideally we'd have used the standard `S` term prefix for
"subject/title" [5], but that prefix is already in use in
`InfoRequestEvent`. We can create our own prefix, but using a standard
would help reduce the complexity of what we give weight to if we applied
weighting to other indexed models:

    # InfoRequestEvent
    acts_as_xapian terms: [ [:request_title, 'S', 'title'] ]

    # User
    acts_as_xapian terms: [ [:name, 'S', 'name'] ]

[1] https://trac.xapian.org/wiki/FAQ/ExtraWeight
[2] https://github.com/mysociety/alaveteli/pull/4544
[3]
https://xapian.org/docs/apidoc/html/classXapian_1_1Document.html#a161943bb133fea0da4a21fa37621764b
[4]
https://xapian.org/docs/apidoc/html/classXapian_1_1Document.html#a730bb59dda98f19c61fe67360f0adb3a
[5] https://xapian.org/docs/omega/termprefixes.html