Deprioritise indexed Xapian texts
This deprioritises the WDF increment for `texts`, mainly so that direct
matches on `PublicBody#name` appear higher in the search results.

The default weight when adding a term is `1` [1], so just _lowering_ the
weight of the `texts` is enough to prioritise `terms`.

Xapian gives the option of doing this either at index time or at search
time [2].

@lizconlan attempted the latter [3], but it resulted in having to jam
lots of extra logic in to our search methods. Given indexing WDTK's
public bodies only takes around 5 minutes (for ~23K records) it seems
more efficient to do this at index time. It also results in much less
code changed.

The actual weighting is quite strange. The example [2] seems to suggest
that you want a lower termpos and a higher WDF increment for a
prioritised term. I'm not entirely sure what the interplay between the
term position and WDF increment is. For future reference, I'm including
links to `add_posting` [4] and `add_term` [1].

Note that the specs assert on `PublicBody#name` instead of comparing
object equality because the rspec diff isn't very helpful now that name
isn't an attribute on `PublicBody` directly (its on
`PublicBody::Translation`).

[1]
https://xapian.org/docs/apidoc/html/classXapian_1_1Document.html#a730bb59dda98f19c61fe67360f0adb3a
[2] https://trac.xapian.org/wiki/FAQ/ExtraWeight
[3] https://github.com/mysociety/alaveteli/pull/4544
[4]
https://xapian.org/docs/apidoc/html/classXapian_1_1Document.html#a161943bb133fea0da4a21fa37621764b