Rescue all errors when indexing
When we hit a record that can't be indexed becauase of an error, we:

1. Get cron output every 5 minutes as the next run of the indexing
   script continuously tries to index the record again.
2. Prevent indexing of other records, because the script is exiting at
   the point of error.

This commit fixes point 2.

We rescue from `StandardError` so that we just output any problems to
STDERR – we'll get a notification message for records that error, but
other records will get indexed with no problem.

We'll continue to receive cron error output every 5 minutes as the
script will run over the job still left in the queue. We can remove the
job from the indexing queue if we want to silence the errors until the
next time it gets marked for reindexing.

Originally I'd resuced from `Exception`, but this is a code smell.
[1][2]

We were debating keeping this as there's not much we don't want to
rescue from, but `StandardError` _seems_ to be high enough in the
hierarchy to catch most things, so lets give it a go like this first.

[1] https://robots.thoughtbot.com/rescue-standarderror-not-exception
[2] https://stackoverflow.com/a/10048406/387558