Prevent continuous retries of failing index jobs
When an indexing job fails, we get an error message via Cron. Before
this commit, we'd continue to try to index the record every time the
update index script was called (every 5 minutes) and therefore receive
an error for the same job every 5 minutes until we fixed the problem.

This commit removes jobs from the queue whatever the result was. We'll
still get the error message, but won't be flooded with output for single
records that can't be indexed.