More error handling in temp:populate_missing_attachment_files
This task kept failing because of unhandled errors. This addresses the
ones that came up. There might be more, but if we rescue from
`StandardError` we might miss the backtrace as it scrolls past, so I
think its better to just exit if we encounter a new type of error,
investigate, and then add more error handling if appropriate.

Related to https://github.com/mysociety/alaveteli/issues/4156.