Fix case sensitivity bug in password reset form
If the user enters their email with a slightly different case than what
we have in the database, the form will silently fail by rendering the
"Now check your email!" template, but without actually sending the mail.

We do this so that you can't probe for registered email accounts, but
obviously in the case where the account exists we should actually send
the reset email.

This commit uses `User.find_user_by_email` which performs a
case-insensitive find on the email address attribute.