Add xapian:restore:queue_events_to_reindex task
This was the first step [1] of restoring from a Xapian backup.

The inline comments describe what its doing fairly well, but the gist is
that we wanted to index anything that had been updated _after_ the date
that the backup was made.

This obviously wants to get refactored out in to a class with specs, but
I'm going to get the tasks in place first before extracting
commonalities.

[1]
https://gist.github.com/garethrees/e9d436f1c506ab801fa622a27917799e#file-reindex_events-rb