Add xapian:restore:queue_obsolete_indexes_for_removal task
This was the second step [2] of restoring from a Xapian backup.

The restored Xapian database won’t have the up-to-date information of
records that have been deleted from the relational DB, so we need to
remove anything that no longer exists from the restored Xapian database.

Note that I used `missing_ids.map` as there’s no `#map!` for a
PG::Result.

This obviously wants to get refactored out in to a class with specs, but
I'm going to get the tasks in place first before extracting
commonalities.

[2]
https://gist.github.com/garethrees/e9d436f1c506ab801fa622a27917799e#file-remove_deleted-rb