Add xapian:restore:second_pass task
This was the third step [1] of restoring from a Xapian backup.

This also includes the `third_pass.rb` script. I don’t think there’s
much benefit of separating the tasks out, so having one less thing to
run is preferable.

This is mostly documented inline, but the gist is that it looks through
records which might need reindexing that wouldn’t have got caught by
reindexing events that were created after the restore.

This obviously wants to get refactored out in to a class with specs, but
I'm going to get the tasks in place first before extracting
commonalities.

[1] https://gist.github.com/garethrees/e9d436f1c506ab801fa622a27917799e#file-second_pass-rb
[2] https://gist.github.com/garethrees/e9d436f1c506ab801fa622a27917799e#file-third_pass-rb