Fix broken migrations
49fd980 and later 2df25d6 introduced migrations with a timestamp two
characters too long. This breaks `rake db:rollback` because these
migrations end up getting applied last – or indeed rolled back instead
of the actually most recent migration.

This commit:

* Fixes migration filenames
* Adds a migration to fix broken timestamps in the schema_migrations
  table

Fixing the filenames is trivial – the files just get renamed removing
the extra digits. Note that in `2017071826152436` we remove `26` rather
than the `36` at the end so that we don't get an invalid time of
`261524`.

The difficulty here is matching up the `schema_migrations` entries with
the new files.

In the case of the `up` migration, we know the old migrations with the
fixed filenames will get applied first. I've added conditionals here so
that if the columns already exist, the only thing that will happen is
that the new `schema_migrations` entry gets recorded. We can then simply
delete the old `version` entry as the file no longer exists, and we know
the migration has been applied.

If we want to rollback from _this_ migration, we only want to revert to
the old entries in `schema_migrations` if the old filename exists. We
need to do this because if we don't, the migration will get re-attempted
on `db:migrate` and fail because the columns already exist. I don't
think this case is likely, but its correct to include this safeguard.

Fixes https://github.com/mysociety/alaveteli/issues/4467.