Make it quicker to ban users in admin interface
Adds a controller to handle account suspensions. It sets a default
suspension message, but the main use right now is to ban for spamming.

I've added buttons to the user list and user show page to quickly set a
"Banned for spamming" message.

This could later be extended with an
`AdminUsersAccountSuspensionsController#destroy` to unban users (reset
`User#ban_text` to `''`), or further buttons could be added with other
common suspension reasons.

This was prompted by an influx of "tech support" spam accounts and
having to manually copy/paste the ban text in each.