Fix Pro batch button alignment bug
CSS styles were still expecting an `input`, but the element changed to
a `button` a long time ago.

Also, `flex-align('flex-start')` was never going to work (it should be
a keyword, rather than a string, for a start…) and aligning to center
actually looks better when the authority name wraps onto multiple lines.

Part of mysociety/alaveteli-professional#502.