Make populate_missing_timestamps more resilient
Sometimes we can have orphaned records, so we hit a NoMethodError error
when trying to calculate the value.

This commit adds `try` calls so that we just revert to `nil` if we
really can't find a value. We'll need to fix these later, but it might
just be that we need to delete an orphaned record rather than actually
calculate a `created_at`/`updated_at` attribute in those cases.