Disable "Publish requests" during Batch sending
If a user publishes requests before an InfoRequestBatch has finished
sending, the existing requests are published but the remaining requests
get created with an Embargo. This means they get hidden from the
now-public InfoRequestBatchController#show action.

I've got a few questions around the UI here – should we just hide the
button, or disable with a message, or a lodaing spinner or what? – but
I'll leave that for an "improvement" ticket rather than worrying about
that in this bugfix.

For this spec we were creating a batch in an unrealistic way – basically
creating the batch without associating it with public bodies and runing
`InfoRequestBatch#create_batch!`.

Now that we prevent Embargo changes when a batch isn't fully sent, this
spec fails as the test data isn't in a state that matches a real-life
sent batch.

This commit does the minimum to fix that. Really we should have a
factory for creating a "real" batch, as well as a stubbed version for
more isolated testing.

Fixes https://github.com/mysociety/alaveteli-professional/issues/512.