Remove "hidden from the site" text from request status messages
This text is misleading as setting the status to 'not_foi' or
'vexatious' does not automatically hide the request, this needs to be
set separately using prominence.

This commit prevents publicly visible requests displaying the message
"this has been hidden", both on the request page itself and in search
results and other listings pages.