Show tooltip on prominence when described_state is not_foi or vexatious
Show a warning tooltip for the prominence field as soon as the
described_state dropdown is changed to 'not_foi' or 'vexatious'.

The tooltip is displayed to the right instead of the default
position of above the control as that might draw it off the top of the
screen and placing it below the control obscures the described_state
field in some browsers (e.g. Safari).

The "manual" trigger option prevents the tooltip disappearing when the
cursor rolls over the prominence dropdown.