Show confirm dialog when trying to save request in an invalid state
In case the tooltip when rolling over the save button doesn't provide
sufficient warning (or editing on a touchscreen where the rollover
condition event might never fire), this provides an "Are you sure?"
dialog before the updated request record is saved when the request has
been altered to have a normal prominence and a described_state of either
'not_foi' or 'vexatious'.

Note that this does not prevent the admin from deliberately choosing to
continue publication of either type of request if they deem it to be in
the public interest to do so. The intention is to reduce continued
publication of such material accidentally.