Move time_travel part of spec to before/after steps
A bit clunky but separates out the setup steps that allow the request
to be evaluated as overdue from the business of classifying the request
and examining the result. Without this, included shared examples will
not receive overdue request data.

Arguably doing this for very_overdue requests is not needed as the next
commit does not add shared examples but altered for consistency.