Move the embargo info message to be visible at all times
Previously the embargo expiry date was hidden behind a click. This makes
it visible at all times so its more obvious that:

* The request is embargoed
* The embargo will expire on the given date

Also tweaks the styling of embargo expiry date.

`--important` gives this too much prominence, so this just sets it to a
normal `<p>` style.

The conditional was added so that in the request page sidebar, you don't
get both cases of the embargo info when the request is public.

With no conditional at all (as per cea8a3b8) when the request is public
you'd see:

    This request will be private on Alaveteli until
    This request is public

Martin initially added an `if`/`else`, preventing this problem but
breaking the new request preview. In that case the "Unless you choose a
privacy option…" string was not shown.

The fix to this is to just wrap the embargo existing case in an `if`.

SIDEBAR

* If the request is embargoed:
  * The emabrgo info is rendered
  * The "public" section is rendered but hidden with CSS

* If the request is public:
  * The embargo info is not rendered
  * The "public" section is rendered without being hidden

NEW REQUEST

Both cases are always rendered, but the JavaScript controlling the
embargo duration select field updates the `display` CSS.

This fix is pretty ugly. I think there was good intent to share the
logic between these embargo forms, but the behaviour is diffent enough
to warrant extraction, as dealing with whether the request has an
embargo or not (or _will_ have an embargo) is all pretty complicated.