Pass in a custom spam scorer config when testing new signups
We need to try to strike a balance between preventing spam signups based
solely on the name and email address - the only information we currently
have at this point - and creating too many false positives.

The mappings for the about_me field have been given a weighting of 0
as it will always be empty at signup.