Mark name and site_name params used in plain/text template as html_safe
This template is only going to be rendered as `text/plain` so malicious
user input (e.g. <script> tags) is not going to be rendered as HTML,
allowing us to treat the text as safe.

We only need to do this because we're going to use these arguments in
translation strings and we don't currently have a way to disable
GetText's HTML escaping of interpolated strings.