Don't show wall items for closed accounts
Don't show wall feed when viewing the anonymised and closed profile of a
User. This makes it much more difficult to associate the closed account
with a real person by scanning through the account's requests and
looking for bad redactions.

The code here is obviously gross – we shouldn't really perform the work
in the first place – but this avoids major refactoring to implement the
change in a nicer way.