Distinguish requests vs batches in pro request list
When an individual request appears below a batch request in the pro
request list, the individual request visually appears to be _part of the
batch_. This is made even more extreme when the batch is public and the
individual request is private, due to the padlock icon indenting the
individual request title.

This commit wraps each item in the list in a bounding box, making it
really clear that items are distinct.

A thicker border-bottom looked messy, as it conflicted with the progress
bar of the batch requests.

Fixes https://github.com/mysociety/alaveteli-professional/issues/497.